# backend/spec/openapi.yaml
openapi: "3.1.1"
info:
  title: My Library API
  description: |
    Book library management system.
    All endpoints return JSON. Some endpoints require authentication
    (Bearer token from `/api/auth/login`).
  version: "0.3.0"
  contact:
    name: My Library
servers:
  - url: http://127.0.0.1:8000
    description: Local development
  - url: https://dingfengbo.top/my-library
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublisherCreation:
      type: object
      required: [name]
      properties:
        name: { type: string }
        intro: { type: string }
        logo: { type: string }
    PublisherUpdate:
      type: object
      properties:
        name: { type: string }
        intro: { type: string }
        logo: { type: string }
    PublisherResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        intro: { type: string, nullable: true }
        logo: { type: string, nullable: true }

    # ── Brand ──
    BrandCreation:
      type: object
      required: [name]
      properties:
        name: { type: string }
        intro: { type: string }
    BrandUpdate:
      type: object
      properties:
        name: { type: string }
        intro: { type: string }
    BrandResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        intro: { type: string, nullable: true }

    # ── Series ──
    BookSeriesCreation:
      type: object
      required: [name]
      properties:
        name: { type: string }
        intro: { type: string }
    BookSeriesUpdate:
      type: object
      properties:
        name: { type: string }
        intro: { type: string }
    BookSeriesResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        intro: { type: string, nullable: true }

    # ── Category ──
...
paths:
...
  /api/publishers:
    get:
      tags: [Publishers]
      summary: List publishers (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: name } }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  publishers:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublisherResponse"
                  total_pages: { type: integer }
                  total_publishers: { type: integer }
    post:
      tags: [Publishers]
      summary: Create a publisher
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherCreation"
      responses:
        "200":
          description: Created publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherResponse"
  /api/publishers/{publisher_id}:
    get:
      tags: [Publishers]
      summary: Get publisher details
      parameters:
        - { name: publisher_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Publisher details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherResponse"
    put:
      tags: [Publishers]
      summary: Update a publisher
      security:
        - BearerAuth: []
      parameters:
        - { name: publisher_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherUpdate"
      responses:
        "200":
          description: Updated publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherResponse"
    delete:
      tags: [Publishers]
      summary: Delete a publisher
      security:
        - BearerAuth: []
      parameters:
        - { name: publisher_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ═══════════════════════════════════════════════
  # Brands
  # ═══════════════════════════════════════════════
  /api/brands:
    get:
      tags: [Brands]
      summary: List brands (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: name } }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  brands:
                    type: array
                    items:
                      $ref: "#/components/schemas/BrandResponse"
                  total_pages: { type: integer }
                  total_brands: { type: integer }
    post:
      tags: [Brands]
      summary: Create a brand
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCreation"
      responses:
        "200":
          description: Created brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
  /api/brands/{brand_id}:
    get:
      tags: [Brands]
      summary: Get brand details
      parameters:
        - { name: brand_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Brand details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
    put:
      tags: [Brands]
      summary: Update a brand
      security:
        - BearerAuth: []
      parameters:
        - { name: brand_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResponse"
    delete:
      tags: [Brands]
      summary: Delete a brand
      security:
        - BearerAuth: []
      parameters:
        - { name: brand_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ═══════════════════════════════════════════════
  # Series
  # ═══════════════════════════════════════════════
  /api/series:
    get:
      tags: [Series]
      summary: List series (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: name } }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookSeriesResponse"
                  total_pages: { type: integer }
                  total_series: { type: integer }
    post:
      tags: [Series]
      summary: Create a series
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookSeriesCreation"
      responses:
        "200":
          description: Created series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookSeriesResponse"
  /api/series/{series_id}:
    get:
      tags: [Series]
      summary: Get series details
      parameters:
        - { name: series_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Series details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookSeriesResponse"
    put:
      tags: [Series]
      summary: Update a series
      security:
        - BearerAuth: []
      parameters:
        - { name: series_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookSeriesUpdate"
      responses:
        "200":
          description: Updated series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookSeriesResponse"
    delete:
      tags: [Series]
      summary: Delete a series
      security:
        - BearerAuth: []
      parameters:
        - { name: series_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ═══════════════════════════════════════════════
  # Categories
  # ═══════════════════════════════════════════════
...